Nueva alerta y subida masica

+++ src/app/pages/facturacion/facturacion.component.ts
@@ -3,7 +3,7 @@ import { MatInput } from '@angular/material/input';
 import { MatPaginator } from '@angular/material/paginator';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { MatTableDataSource } from '@angular/material/table';
-import { ToastrService } from 'ngx-toastr';
+import { NotifierService } from 'angular-notifier';
 import { Guia } from 'src/app/Interfaces/Guia';
 import { CheckInternetService } from 'src/app/services/check-internet.service';
 import { GuiaService } from 'src/app/services/guia.service';
@@ -21,12 +21,12 @@ export class FacturacionComponent implements AfterViewInit, OnInit {
   value: string = "";
   isOnline: boolean = true;
   // DataTables Variables
-  displayedColumns: string[] = ['position', 'nroGuia', 'fechaInicio', 'acciones'];
+  displayedColumns: string[] = ['position', 'nroGuia', 'fechaInicio', 'estado', 'acciones'];
   dataSource: MatTableDataSource<Guia>;
   @ViewChild(MatPaginator) paginator: MatPaginator;
 
   constructor(private _guiaService: GuiaService,
-    private toastr: ToastrService,
+    private notifier: NotifierService,
     private _snackBar: MatSnackBar,
     private _loginService: LoginService,
     private _internet: CheckInternetService,
@@ -34,7 +34,7 @@ export class FacturacionComponent implements AfterViewInit, OnInit {
 
   ngOnInit(): void {
     this.checkStatus();
-    if(this.isOnline){
+    if (this.isOnline) {
       this.getData();
     } else {
       this.loadDatasource(this._localStorage.guias);
@@ -45,11 +45,16 @@ export class FacturacionComponent implements AfterViewInit, OnInit {
     this.setFocus();
   }
 
-  checkStatus(){
+  checkStatus() {
     this._internet.createOnline$().subscribe(status => {
       this.isOnline = status;
-      if(this.isOnline){
-        this.getData();
+      if (this.isOnline) {
+        this._guiaService.cargar()?.subscribe(response => {
+          this._localStorage.removeAll();
+          this.getData();
+        });
+      } else {
+        this.getDataFromLS();
       }
     });
   }
@@ -57,16 +62,21 @@ export class FacturacionComponent implements AfterViewInit, OnInit {
   getData() {
     let idlocalidad = this._loginService.localidad;
     this._guiaService.findAll(idlocalidad).subscribe(response => {
-      this.loadDatasource([...response, ...this._localStorage.guias]);
+      this.loadDatasource(response);
     }, err => {
       alert('Ocurrio un error en la obtención de los datos.');
     });
   }
 
-  loadDatasource(data: any){
+  getDataFromLS() {
+    let guias = this._localStorage.guias;
+    this.loadDatasource(guias);
+  }
+
+  loadDatasource(data: any) {
     this.dataSource = new MatTableDataSource<Guia>(data);
-      this.dataSource.paginator = this.paginator;
-      this.paginator._intl.itemsPerPageLabel = "Guías por página";
+    this.dataSource.paginator = this.paginator;
+    this.paginator._intl.itemsPerPageLabel = "Guías por página";
   }
 
   setFocus() {
@@ -80,40 +90,49 @@ export class FacturacionComponent implements AfterViewInit, OnInit {
     const value = this.input.value;
     if (this.isValid(value)) {
       let idlocalidad = this._loginService.localidad;
-      const guia: Guia = { nroguia: value, fechainicio: new Date(), idlocalidad};
-      if(!this.isOnline){
-        this._localStorage.add(guia);
-        this.toastr.success(`Guía ${guia.nroguia} registrada.`);
-        this.dataSource.data = [...this.dataSource.data, guia];
-        this.clear();
+      let guia: Guia = { nroguia: value, fechainicio: new Date(), idlocalidad };
+      if (!this.isOnline) {
+        guia.local = true;
+        if (this._localStorage.add(guia)) {
+          this.notifier.notify('success', `Guía ${guia.nroguia} registrada.`);
+          this.dataSource.data = [...this.dataSource.data, guia];
+          this.clear();
+        } else {
+          this.notifier.notify('error', 'El N° de Guía ya existe.');
+          this.clear();
+        }
       } else {
         this._guiaService.save(guia).subscribe(response => {
           this.dataSource.data = [...this.dataSource.data, response];
           this.clear();
-          this.toastr.success(`Guía ${guia.nroguia} registrada.`, 'Tecnimotors');
+          this.notifier.notify('success', `Guía ${guia.nroguia} registrada.`);
         }, err => {
-          this.toastr.error(err.error, 'Cuidado!!!');
+          this.notifier.notify('error', err.error);
           this.clear();
         });
       }
     } else {
-      this.toastr.info(`N° de Guía inválida.`, 'Tecnimotors');
+      this.notifier.notify('warning', 'N° de Guía inválida');
     }
   }
 
   delete(guia: Guia) {
     if (window.confirm(`¿Estas seguro de eliminar la guía ${guia.nroguia}?`)) {
-      this._guiaService.delete(guia.idproceso).subscribe(response => {
-        this.dataSource.data = this.dataSource.data.filter((objGuia: Guia) => objGuia.idproceso != guia.idproceso);
-        this._snackBar.open(`Guía N° ${guia.nroguia} eliminada.`, "Ok", { duration: 2000 })
-      }, err => {
-        this.toastr.error('Ocurrió un error en la eliminación.');
-      });
+      if (this.isOnline) {
+        this._guiaService.delete(guia.idproceso).subscribe(response => {
+          this.dataSource.data = this.dataSource.data.filter((objGuia: Guia) => objGuia.idproceso != guia.idproceso);
+          this._snackBar.open(`Guía N° ${guia.nroguia} eliminada.`, "Ok", { duration: 2000 })
+        }, err => {
+          this.notifier.notify('error', 'Ocurrió un error en la eliminación.');
+        });
+      } else {
+
+      }
     }
   }
-  
+
   isValid(val: string) {
-    return  val.length !== 0;
+    return val.length !== 0;
   }
 
   clear() {
